U1-M4-Summative added

+++ U1-M2-Summative-Cimafranca-Daryl/CalculatorObject/src/main/java/com/trilogyed/Calculator.java
@@ -26,7 +26,7 @@ public class Calculator {
         return a*b;
     }
 
-    public int divide(int a, int b){
+    public double divide(int a, int b){
         return a/b;
     }
